Moved validation to attribute on model
