+++ UID.Participant.Api.Test/UID.Participant.Api.Test.csproj
@@ -48,8 +48,4 @@
     <Using Include="Xunit" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="ValidationTests\" />
-  </ItemGroup>
-
 </Project>
