+++ UID.Participant.Api.Test/ValidationTests/ClientTypesTests.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using UID.Participant.Api.Test.ControllerTests;
+
+namespace UID.Participant.Api.Test.ValidationTests
+{
+    public class ClientTypesTests : ControllerTestsBase
+    {
+        [Fact]
+        public void UnknowClientTypeReturnsError()
+        {
+
+        }
+    }
+}
